// src/api/models/getSalesStatByPassTypeResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { PassTypeSalesStatItem } from './passTypeSalesStatItem';

export interface GetSalesStatByPassTypeResponse {
  /** 결과 여부 */
  ok: boolean;
  /** 응답 데이터 */
  data: PassTypeSalesStatItem[];
}