// src/api/models/passTypeSalesStatItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Partners
 * Allta Partners 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */

export interface PassTypeSalesStatItem {
  /** 날짜 */
  date: string;
  /** 전체 매출 */
  totalSales: number;
  /** 구독권 매출 */
  subscriptionSales: number;
  /** 일회권 매출 */
  ticketSales: number;
  /** 현장결제 매출 */
  offlineTicketSales: number;
}